insights_api/requirements.txt
Django==1.9.6
argparse==1.2.1
django-filter==0.13.0
djangorestframework==3.3.3
psycopg2==2.6.1
wsgiref==0.1.2
python-dateutil
supervisor-stdout
